pyproject.toml
[project]
name = "map-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "altair>=5.5.0",
    "fastexcel>=0.13.0",
    "folium>=0.19.2",
    "geopandas>=1.0.1",
    "geopy>=2.4.1",
    "google-maps-places>=0.1.23",
    "googlemaps>=4.10.0",
    "ipython>=8.30.0",
    "jupyterlab>=4.3.3",
    "matplotlib>=3.10.0",
    "numpy>=2.2.0",
    "polars>=1.17.1",
    "pyarrow>=19.0.1",
    "pyqt5>=5.15.11",
    "rasterio>=1.4.3",
    "tqdm>=4.67.1",
    "vega-datasets>=0.9.0",
    "xlsxwriter>=3.2.0",
]
